fast-api-instrumentation-v2/requirements.txt
fastapi
fastapi[standard]
uvicorn
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
locust
